refactor: Topbar component 파일위치변경
