+++ apps/web/src/app/(auth)/signin/page.tsx
@@ -5,7 +5,7 @@ import { CategoryGithubIcon, StaticLogoSymbolIcon } from '@80000coding/web-icons
 import { useRouter } from 'next/navigation'
 import { useCallback } from 'react'
 
-import TopbarWithXIcon from '@/components/common/TopbarWithXIcon'
+import TopbarWithXIcon from '@/components/shared/TopbarWithXIcon'
 
 export default function SigninPage() {
   const router = useRouter()
